server/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
sqlalchemy
alembic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]

